requirements.txt
# Core Libraries
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Streamlit App
streamlit==1.34.0
plotly==5.21.0

# NLP Models & Transformers
transformers==4.41.1
sentence-transformers==2.6.1
bertopic==0.16.0
faiss-cpu==1.11.0

# Sentiment & Text Processing
textblob==0.18.0
nltk==3.8.1

# GPT Agent + LangChain
openai==1.30.3
langchain==0.1.20
tiktoken==0.6.0

# HTML & PDF Reporting
jinja2==3.1.4
weasyprint==60.2  # Note: Requires system dependencies (GTK, Cairo)
pdfkit==1.0.0

# .env configuration
python-dotenv==1.0.1
google-generativeai==0.3.2
